refactor: restructure packages
