+++ elasticai/creator/file_generation/on_disk_path.py
@@ -1,8 +1,8 @@
 import os
 from pathlib import Path as _PyPath
 
-from elasticai.creator.file_generation.savable import File, Path
-from elasticai.creator.templating.template import Template, TemplateExpander
+from .savable import File, Path
+from .template import Template, TemplateExpander
 
 
 class OnDiskFile(File):
