+++ elasticai/creator/nn/conv1d/design.py
@@ -1,7 +1,10 @@
 import math
 
 from elasticai.creator.file_generation.savable import Path
-from elasticai.creator.templating.template import InProjectTemplate, module_to_package
+from elasticai.creator.file_generation.template import (
+    InProjectTemplate,
+    module_to_package,
+)
 from elasticai.creator.vhdl.auto_wire_protocols.port_definitions import create_port
 from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.design.ports import Port
