+++ elasticai/creator/vhdl/base_template_generator/base_template_generator.py
@@ -1,4 +1,4 @@
-from elasticai.creator.templating.template import (
+from elasticai.creator.file_generation.template import (
     InProjectTemplate,
     TemplateExpander,
     module_to_package,
